backend/requirements.txt
Django==5.0.1
python-dotenv==1.0.0
mysqlclient==2.2.4
djangorestframework==3.14.0